requirements.txt
openai
langchain
langchain-openai
langchain-community
langchain-chroma
langgraph
fastapi
uvicorn
python-dotenv
chromadb
requests
tavily-python
sqlalchemy
psycopg2-binary
celery
redis
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
slack-sdk
streamlit
alembic
pydantic